refactor(frontend): reorganize src structure
