+++ packages/frontend/src/index.html
@@ -4,7 +4,7 @@
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<title>Org Roam UI Lite</title>
-		<link rel="stylesheet" href="./app.css" />
+                <link rel="stylesheet" href="./assets/app.css" />
 		<script type="module" src="./main.ts"></script>
 	</head>
 	<body>
